Made some slight edits to the eslint file.
Moving files around, reorganizing structure of app. Creating new folders and separate components for each jsx file. All turning into stateless / ES6.
