+++ app/components/HomePage/BodyButton.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Link } from 'react-router';
+
+const BodyButton = (props) => {
+
+		this.displayName = 'BodyButton';
+	render() {
+		return (
+			<Link to="/signup" role="button" className="button-home">
+				JOIN US
+			</Link>
+		);
+	}
+}
+
+module.exports = BodyButton;
