+++ app/components/HomePage/DriverDescription.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const DriverDescription = () => (
+	<div className="source-sans" style={{ marginBottom: 40 }}>
+		<p>There are people who would rather see their extra food do some good instead of going to waste! We accept donations from business such as catering companies & facilites, restaurants and forcery stores, as well as donations from individuals.</p>
+		<p>As an extra benefit to all of our donors, we provide tax receipts to them so they can write their donations off when they do their taxes.</p>
+		<p>Food Drivr provides an easy way for both drivers and donors to help put an end to hunger in their communities.</p>
+	</div>
+);
+
+module.exports = DriverDescription;
