+++ app/components/HomePage/SectionComments.jsx
@@ -1,0 +1,48 @@
+import React from 'react';
+
+class SectionComments extends React.Component {
+	constructor(){
+		this.state = {
+			text: "We're so happy you came in! We've been throwing away hundreds of cupcakes and it just makes us sick. Thank you for what you're doing.",
+			author: 'CARLOS ESTRADA',
+			company: 'Jones Brothers Cupcakes',
+			item: 0
+		};
+	}
+
+	selectItem(value) {
+		this.setState({ item: value });
+		if (value === 0) {
+			this.setState({ text: "We're so happy you came in! We've been throwing away hundreds of cupcakes and it just makes us sick. Thank you for what you're doing.",
+							author: 'CARLOS ESTRAD',
+							company: 'Jones Brothers Cupcakes' });
+		} else if (value === 1) {
+			this.setState({ text: 'Some more text about someone from a random company 1 should go here!',
+							author: 'NAME LASTNAME 1',
+							company: 'Super Company 1' });
+		} else {
+			this.setState({ text: 'A different kind of text about someone from a random company 2 should go here!',
+							author: 'NAME LASTNAME 2',
+							company: 'Super Company 2' });
+		}
+	}
+
+	isActive(value) {
+		return ((this.state.item === value) ? 'active-bullet ' : '') + 'bullet pointer-cursor';
+	}
+
+	render() {
+		return (
+			<div className="homepage-comments text-center text-white">
+				<Comment text={this.state.text} author={this.state.author} company={this.state.company} />
+				<div className="comments-dots">
+					<div className={this.isActive(0)} onClick={this.selectItem.bind(this,0)} />
+					<div className={this.isActive(1)} onClick={this.selectItem.bind(this,1)} />
+					<div className={this.isActive(2)} onClick={this.selectItem.bind(this,2)} />
+				</div>
+			</div>
+		);
+	}
+});
+
+module.exports = SectionComments;
