+++ app/components/HomePage/VolunteerDescription.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const VolunteerDescription = () => (
+	<div className="source-sans" style={{ marginBottom: 40 }}>
+		<p>There are people who would rather see their extra food do some good instead of going to waste! We accept donations from business such as catering companies & facilites, restaurants and forcery stores, as well as donations from individuals.</p>
+		<p>Some other text for volunteers!</p>
+	</div>
+);
+
+module.exports = VolunteerDescription;
