+++ app/components/Reusable/TruckButton.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const TruckButton = () => <div className="truck-button"></div>;
+
+module.exports = TruckButton;
