+++ app/components/Reusable/WhiteTruckButton.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Link } from 'react-router';
+
+const WhiteTruckButton = () => (
+	<Link to="/">
+		<div className="white-truck-button"></div>
+	</Link>
+);
+
+module.exports = WhiteTruckButton;
