+++ app/config/routes.js
@@ -1,8 +1,5 @@
-/*Dependencies*/
 import React from 'react';
-import ReactDOM from 'react-dom';
 import { browserHistory, Router, Route, IndexRoute } from 'react-router';
-/*Components*/
 import App from '../components/main.js';
 import HomePage from '../components/homepage-components.jsx';
 import SignInPage from '../components/signin-components.jsx';
@@ -12,8 +9,7 @@ import ThankYou from '../components/thank-you-components.jsx';
 import Registration from '../containers/registration-container.jsx';
 import auth from '../utils/auth.js';
 import injectTapEventPlugin from 'react-tap-event-plugin';
-import * as Pages from 'pages';
-import foodDrivrAPI from '../utils/foodDrivrAPI';
+import * as Pages from '../pages';
 injectTapEventPlugin();
 
 function requireAuth(nextState, replace) {
@@ -21,19 +17,19 @@ function requireAuth(nextState, replace) {
 		replace({
 			pathname: '/',
 			state: { nextPathname: nextState.location.pathname }
-		})
+		});
 	}
 }
 
-var routes = (
+const routes = (
 	<Router history={browserHistory}>
 		<Route path="/" component={App}>
 			<IndexRoute component={HomePage} />
 			<Route path="signin" component={SignInPage} />
 			<Route path="signup" component={SignUpPage} />
-			<Route path="signup/donor" component={Registration} header="Donor"/>
-			<Route path="signup/volunteer" component={Registration} header="Volunteer"/>
-			<Route path="donation" component={DonationList} onEnter={requireAuth}/>
+			<Route path="signup/donor" component={Registration} header="Donor" />
+			<Route path="signup/volunteer" component={Registration} header="Volunteer" />
+			<Route path="donation" component={DonationList} onEnter={requireAuth} />
 			<Route path="thankyou" component={ThankYou} />
 			<Route path="profile" component={Pages.UserProfilePage} onEnter={requireAuth} />
 			<Route path="thankyou/:userType" component={ThankYou} />
